refactor players and entities into header files
