+++ src/main.cpp
@@ -1,31 +1,11 @@
 #include <../include/raylib.h>
 #include <../include/defs.h>
 #include <../include/entity.h>
+#include <../include/camera.h>
+#include <../include/player.h>
 
 #include <stdio.h>
 #include <cmath>
-
-
-
-
-struct Viewport {
-    Camera2D camera;
-    Entity* parent;
-
-    void init(Entity* following) {
-        camera = {0};
-        
-        camera.offset = { SCREEN_W / 2.f, SCREEN_H / 2.f };
-        camera.rotation = 0.f;
-        camera.zoom = 1.f;
-
-        parent = following;
-    }
-
-    void update() {
-        camera.target = parent->position;
-    }
-};
 
 int main() {
     SetConfigFlags(FLAG_VSYNC_HINT);
@@ -39,6 +19,8 @@ int main() {
 
     while (!WindowShouldClose()) {
         float dt = GetFrameTime();
+
+        playerInputHandle(&player, dt);
 
         // UPDATE
         player.update(dt);
